src/parcels/requirements.txt
boto3==1.16.2
botocore==1.19.2
certifi==2020.12.5
chardet==4.0.0
idna==2.10
jmespath==0.10.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.25.1
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.11
